Minor refactor and fix build. link #125
