+++ app/src/main/java/com/cryart/sabbathschool/ui/login/LoginActivity.kt
@@ -36,7 +36,6 @@ import com.cryart.sabbathschool.databinding.SsLoginActivityBinding
 import com.cryart.sabbathschool.databinding.SsLoginButtonsBinding
 import com.cryart.sabbathschool.lessons.ui.quarterlies.QuarterliesActivity
 import com.cryart.sabbathschool.ui.LessonsActivity
-import com.cryart.sabbathschool.ui.MainActivity
 import com.facebook.CallbackManager
 import com.facebook.login.LoginManager
 import com.google.android.gms.auth.api.signin.GoogleSignIn
