+++ features/lessons/src/main/java/com/cryart/sabbathschool/ui/LessonItemViewHolder.kt
@@ -1,6 +1,6 @@
 package com.cryart.sabbathschool.ui
 
 import androidx.recyclerview.widget.RecyclerView
-import com.cryart.sabbathschool.databinding.ItemLessonBinding
+import com.cryart.sabbathschool.lessons.databinding.ItemLessonBinding
 
 class LessonItemViewHolder(binding: ItemLessonBinding) : RecyclerView.ViewHolder(binding.root)
